3D Volume
This calculates the height, length, radius, ect. inputted by the user to different 3D shapes.

+++ 3DVolume.py
@@ -4,7 +4,7 @@
 #Cube
 
 #This asks the user to input the lengths for the cube and calculates it
-length=int(input("Please enter the length of one edge of your cube:"))
+length=float(input("Please enter the length of one edge of your cube:"))
 cube=length*length*length
 #This tells the user the volume of the cube in units
 print("The cube's volume is:", cube," units")
@@ -12,7 +12,7 @@ print("The cube's volume is:", cube," units")
 #Sphere
 
 #This asks the user to input the radius for the sphere and calculates it
-radius=int(input("Please enter the radius of your sphere:"))
+radius=float(input("Please enter the radius of your sphere:"))
 sphere=4/3*3.1416*radius*radius*radius
 #This tells the user the volume of the sphere in units
 print("The sphere's volume is:", sphere, " units")
@@ -20,8 +20,8 @@ print("The sphere's volume is:", sphere, " units")
 #Cone
 
 #This asks the user to input the height and radius for the cone and calculates it
-height=int(input("Please enter the height of your cone:"))
-radius=int(input("Please enter the radius of your cone:"))
+height=float(input("Please enter the height of your cone:"))
+radius=float(input("Please enter the radius of your cone:"))
 cone=1/3*3.1416*radius*radius*height
 #This tells the user the volume of the sphere in units
 print("The cone's volume is:", cone, "units")
@@ -29,8 +29,8 @@ print("The cone's volume is:", cone, "units")
 #Cylinder
 
 #This asks the user to input the height and radius for the cylinder and calculates it
-height=int(input("Please enter the height of your cylinder:"))
-radius=int(input("Please enter the radius of your cylinder:"))
+height=float(input("Please enter the height of your cylinder:"))
+radius=float(input("Please enter the radius of your cylinder:"))
 cylinder=3.1416*radius*radius*height
 #This tells the user the volume of the cylinder in units
 print("The cylinder's volume is:", cylinder, "units")
